+++ cmd/livesim2/app/handler_urlgen.go
@@ -191,9 +191,10 @@ type nameWithSelect struct {
 type segmentTimelineType string
 
 const (
-	Number         segmentTimelineType = "nr"
-	TimelineTime   segmentTimelineType = "tlt"
-	TimelineNumber segmentTimelineType = "tlnr"
+	Number              segmentTimelineType = "nr"
+	TimelineTime        segmentTimelineType = "tlt"
+	TimelineNumber      segmentTimelineType = "tlnr"
+	TimelineTimePattern segmentTimelineType = "tltp"
 )
 
 func (s *Server) createInitData(aInfo assetsInfo) (data urlGenData, err error) {
@@ -249,6 +250,9 @@ func createURL(r *http.Request, aInfo assetsInfo, drmCfg *drm.DrmConfig) urlGenD
 	case TimelineNumber:
 		data.Stl = TimelineNumber
 		sb.WriteString("segtimelinenr_1/")
+	case TimelineTimePattern:
+		data.Stl = TimelineTimePattern
+		sb.WriteString("segtimeline_pattern/")
 	default:
 		fmt.Printf("Bad stl: %s\n", stl)
 	}